Add Flask application with prediction functionality and configuration

+++ src/pipeline/predict_pipeline.py
@@ -34,8 +34,8 @@ class CustomData:
                  parental_level_of_education: str,
                  lunch: str,
                  test_preparation_course: str,
-                 reading_score: float,
-                 writing_score: float):
+                 reading_score: int,
+                 writing_score: int):
         self.gender = gender
         self.race_ethnicity = race_ethnicity
         self.parental_level_of_education = parental_level_of_education
